fix: decimal balance

+++ saleor/graphql/account/types.py
@@ -307,7 +307,7 @@ class User(ModelObjectType[models.User]):
     user_type = graphene.String(
         required=True, description="The type of the user. Defined by the OIDC provider"
     )
-    balance = graphene.Int(required=True, description="The balance of the user.")
+    balance = graphene.Float(required=True, description="The balance of the user.")
     code = graphene.String(required=True, description="The code of a user.")
     continuous = graphene.Int(
         required=True, description="The continous login days of the user."
